requirements.txt
bert-serving-client
numpy
torch
csv